Add profile image to header and setup VIM
I do not think this will be lead to any changes in the project files,
but VIM is now setup both within VSC and within git for the use of
editing commit messages. I may need to change it in the future so
that it opens in VSC instead...

I'm now seeing that I definitely need to make commit messages open
in VSC, so I will work on that between now and the next committed file.